refactor(restir-di): ResB para UINT e MCap relativo como o paper
Fase 2 do projeto de mesh lights, e ela conserta de passagem dois tetos que ja doiam hoje.

O ResB era RGBA16F com o LightIndex guardado como float. Inteiro em fp16 e exato so ate 2048, e
o levantamento da cena de teste deu 26.498 triangulos emissivos: mesh lights quebrariam no
primeiro frame, sombreando a luz errada em silencio. M e idade tambem dividiam um unico canal,
o que prendia M em 63 e impedia o MCap relativo do paper.

Novo layout, R32G32B32A32_UINT:
  .x = LightIndex, 32 bits (0xFFFFFFFF = sem amostra)
  .y = RESERVADO p/ a UV da amostra na superficie da luz (16+16). Um triangulo emissivo nao e
       identificado so pelo indice; deixar o campo pronto evita repacotar duas vezes. Hoje 0.
  .z = M (16 bits) | idade (8 bits) | livre (8)
  .w = normal do ponto visivel, octaedrica em unorm 16+16

Com M em 16 bits o MCap vira RELATIVO, como o paper manda ("clamp do M do frame anterior a no
maximo 20x o M do reservoir do frame atual"). Era 20 ABSOLUTO, ou seja ~8x menos historico do que
a referencia pretende: com 8 candidatas o teto correto e 160. A multiplicacao mora na CPU de
proposito — o shader continua lendo Sampling.y como teto absoluto, entao nao ha como um passe
capar em 20 e o outro em 160.

O MaxAge tambem deixa de ter restricao de formato: a idade agora vai a 255. Fica em 12 mesmo
assim, porque subir tem que ser medido e nao presumido.

Custo: 8 B/pixel a mais por buffer, ~16 MB a 1080p somando o ping-pong.

+++ Shaders/Lighting/ReSTIRDISpatial.cs.hlsl
@@ -32,7 +32,7 @@ Texture2D<float>  Depth    : register(t3);
 RaytracingAccelerationStructure Scene : register(t4);
 StructuredBuffer<InstanceGeo> Instances : register(t5);
 Texture2D<float4> ResA : register(t6);
-Texture2D<float4> ResB : register(t7);
+Texture2D<uint4>  ResB : register(t7);
 StructuredBuffer<FGPULightFull> Lights : register(t8);
 #include "../Temporal/TemporalMotionCommon.hlsli"
 StructuredBuffer<FTemporalInstanceTransform> TemporalTransforms : register(t9);
